添加本地账单

+++ app/src/main/java/com/copasso/cocobill/mvp/presenter/Imp/MonthDetailPresenterImp.java
@@ -43,7 +43,7 @@ public class MonthDetailPresenterImp extends MonthDetailPresenter implements Mon
     }
 
     @Override
-    public void deleteBill(int id) {
+    public void deleteBill(Long id) {
         monthDetailModel.delete(id);
     }
